Separate HTTP from Event constants
