+++ src/main/java/org/gridsuite/study/server/StudyConstants.java
@@ -63,10 +63,6 @@ public final class StudyConstants {
     public static final String QUERY_PARAM_REPORT_UUID = "reportUuid";
     public static final String QUERY_PARAM_REPORTER_ID = "reporterId";
     public static final String QUERY_PARAM_REPORT_TYPE = "reportType";
-    public static final String HEADER_RECEIVER = "receiver";
-    public static final String HEADER_BUS_ID = "busId";
-    public static final String HEADER_IMPORT_PARAMETERS = "importParameters";
-    public static final String HEADER_MESSAGE = "message";
     public static final String HEADER_USER_ID = "userId";
     public static final String QUERY_PARAM_ONLY_STASHED = "onlyStashed";
     public static final String QUERY_PARAM_STASHED = "stashed";
